third_party/WebKit/LayoutTests/http/tests/workers/shared-worker-importScripts.html

<!DOCTYPE html>
<title>SharedWorker: importScripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
</script>

<script>
promise_test(t => {
  const worker = new SharedWorker('resources/worker-importScripts.js', 'importScripts');
  return new Promise(resolve => worker.port.onmessage = resolve)
      .then(evt => {
        assert_equals(evt.data, 'DxONE');
      });
}, 'foo');
</script>
